requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==4.0.1
django-celery-results==2.2.0
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.10.1
kombu==5.2.3
Markdown==3.3.6
prompt-toolkit==3.0.24
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2021.3
redis==4.1.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0
